+++ packages/genesys/packages/react-components-core/src/SelectGroup/Selector/Selector.types.ts
@@ -1,0 +1,13 @@
+import { ReactNode } from "react";
+import { SelectorChildrenContext } from "../SelectGroup.types";
+
+export interface SelectorProps<T> {
+    /**
+     * Selector value
+     */
+    value: T;
+    /**
+     * Selector content
+     */
+    children: (context: SelectorChildrenContext<T>) => ReactNode;
+}