+++ packages/genesys/packages/react-components-core/src/SelectGroup/index.tsx
@@ -1,0 +1,3 @@
+export * from "./Selector";
+export * from "./SelectGroup.types";
+export * from "./SelectGroupContext";