+++ packages/genesys/packages/react-components/src/SelectGroup/Selector/Selector.types.ts
@@ -1,0 +1,14 @@
+import { CoreSelectorProps } from "@peersyst/react-components-core";
+import { CSSProperties } from "react";
+import { LabelProps } from "../../Label";
+
+export interface SelectorProps<T> extends CoreSelectorProps<T, LabelProps> {
+    /**
+     * Selector className
+     */
+    className?: string;
+    /**
+     * Selector style
+     */
+    style?: CSSProperties;
+}