+++ packages/genesys/packages/react-components/src/SelectGroup/index.tsx
@@ -1,0 +1,3 @@
+export * from "./Selector";
+export { default as SelectGroup } from "./SelectGroup";
+export * from "./SelectGroup.types";