+++ packages/genesys/packages/react-components/src/config/components.config.tsx
@@ -302,11 +302,21 @@ const componentsConfig: ComponentsConfig = {
     Select: {
         defaultProps: {},
     },
+    SelectGroup: {
+        defaultProps: {},
+    },
     SelectItem: {
         defaultProps: {},
     },
     SelectMenu: {
         defaultProps: {},
+    },
+    Selector: {
+        defaultProps: {
+            LabelProps: {
+                placement: "right",
+            },
+        },
     },
     Skeleton: {
         defaultProps: {