+++ packages/genesys/packages/react-components/src/config/components.config.types.ts
@@ -60,6 +60,8 @@ import { BlockchainAddressProps } from "../BlockchainAddress";
 import { ChipProps } from "../Chip";
 import { AlertProps } from "../Alert";
 import { ColorInputProps } from "../ColorInput";
+import { SelectorProps } from "../SelectGroup";
+import { SelectGroupProps } from "../SelectGroup/SelectGroup.types";
 
 /**
  * No config for:
@@ -122,8 +124,10 @@ export interface RadioButtonConfig extends ComponentConfig<RadioButtonProps> {}
 export interface RangeSliderConfig extends ComponentConfig<RangeSliderProps> {}
 export interface RowConfig extends ComponentConfig<RowProps> {}
 export interface SelectConfig extends ComponentConfig<SelectProps<any>> {}
+export interface SelectGroupConfig extends ComponentConfig<SelectGroupProps<any>> {}
 export interface SelectItemConfig extends ComponentConfig<SelectItemProps<any>> {}
 export interface SelectMenuConfig extends ComponentConfig<SelectMenuProps> {}
+export interface SelectorConfig extends ComponentConfig<SelectorProps<any>> {}
 export interface SkeletonConfig extends ComponentConfig<SkeletonProps> {}
 export interface SliderConfig extends ComponentConfig<SliderProps> {}
 export interface SwitchConfig extends ComponentConfig<SwitchProps> {}
@@ -189,8 +193,10 @@ export interface ComponentsConfig {
     RangeSlider: RangeSliderConfig;
     Row: RowConfig;
     Select: SelectConfig;
+    SelectGroup: SelectGroupConfig;
     SelectItem: SelectItemConfig;
     SelectMenu: SelectMenuConfig;
+    Selector: SelectorConfig;
     Skeleton: SkeletonConfig;
     Slider: SliderConfig;
     Switch: SwitchConfig;