+++ packages/genesys/packages/react-components/src/index.tsx
@@ -45,6 +45,7 @@ export * from "./ProgressBar";
 export * from "./RadioButton";
 export * from "./Row";
 export * from "./Select";
+export * from "./SelectGroup";
 export * from "./Skeleton";
 export * from "./Slider";
 export * from "./SvgIcon";